refactor(stelae): move the backfill daemon into the profile crate
`snapshot backfill`'s 1,122-line publisher daemon lived in the CLI, where
nothing outside a `dolos` binary could reach it and nothing tested it but
its own arithmetic. It moves to `dolos-snapshot` behind a new `mithril`
feature, generic over `dolos_core::Domain`, together with the mithril
fetch it drives.

What moved:

- `crates/snapshot/src/mithril.rs` — the aggregator client, the download
  plan and the verification, shared by the daemon and `bootstrap mithril`.
  The feedback receiver is now a parameter rather than a TUI the fetch
  builds for itself, so the library draws nothing.
- `crates/snapshot/src/backfill.rs` — the `Step`/`Advance`/`Import` loop,
  the window arithmetic, the cleanup and the import, with their tests.
  The WAL reseed goes through `dolos_core::seed_wal_from_state`.

What stayed in `src/bin/dolos/snapshot/backfill.rs`: clap args, the tokio
runtime, the SIGTERM/SIGINT watcher, the progress bars, and the seams a
`Domain` does not expose — opening the stores without a domain, building
and tearing down one with a forced `stop_epoch`, and the publish, which
stays CLI-side so `snapshot publish --repo` and the daemon keep one
telling of the repository arm and one report. `to_repository` now takes
the three stores it reads rather than the binary's own store set.

Runtime and signal ownership stay with the binary, which is what keeps
the shutdown semantics unchanged.

Verified: `cargo test --workspace --all-targets`, clippy `--all-features`
`-D warnings`, `cargo +nightly-2026-08-27 fmt --all --check`, goldens with
zero re-pins, and `cargo check -p dolos-snapshot --all-targets` under
`oci`, `mithril` and `oci,mithril` — the combinations CI now covers. A
preview backfill was run end to end against a local `registry:2`: SIGTERM
mid-window stopped on the interrupted-run message with the stores intact,
and the rerun resumed, replayed to the epoch-1 boundary, pruned 54
consumed immutable files and published sequence 1.

Plan: plans/dolos-stelae-repo-split-backfill-daemon.md
