+++ crates/snapshot/Cargo.toml
@@ -12,6 +12,19 @@ edition.workspace = true
 # place. `dolos`'s own `registry` feature is what turns this on.
 oci = ["stelae/oci", "stelae-driver/oci"]
 
+# The mithril half of the publisher: the aggregator fetch (`src/mithril.rs`)
+# and the backfill daemon that drives it (`src/backfill.rs`). Default-off, and
+# the only reason this crate ever pulls `mithril-client`, pallas or a tokio
+# runtime handle — a build that only writes steles keeps the dependency tree it
+# had. `dolos`'s own `mithril` feature is what turns this on.
+mithril = [
+    "dep:mithril-client",
+    "dep:pallas",
+    "dep:itertools",
+    "dep:tokio",
+    "dep:tokio-util",
+]
+
 # This crate is the *profile* half of the Stelae boundary: it may depend on
 # `stelae` and on `dolos-*`, never the other way around. See
 # adrs/004_stelae_snapshots.md, "Code layout".
@@ -27,6 +40,20 @@ serde.workspace = true
 serde_json.workspace = true
 thiserror.workspace = true
 tracing.workspace = true
+
+# Feature `mithril` only. `rustls-no-provider` keeps reqwest's hard-wired
+# aws-lc-rs crypto provider out of the build; the program installs ring as the
+# process-default provider instead — `dolos`'s `main()`, and the fixture under
+# `cargo test`.
+mithril-client = { version = "0.14.5", optional = true, default-features = false, features = [
+    "fs",
+    "num-integer-backend",
+    "rustls-no-provider",
+] }
+pallas = { workspace = true, optional = true }
+itertools = { workspace = true, optional = true }
+tokio = { workspace = true, optional = true, features = ["macros"] }
+tokio-util = { workspace = true, optional = true }
 
 [dev-dependencies]
 tempfile = "3.20.0"
